+++ product-service/src/utils/validation/validate.js
